requirements.txt
django
faiss-cpu
sentence-transformers
numpy
pandas
